break macro logic into its own file
